Move jobs API to dedicated directory
