+++ src/core/index.ts
@@ -1,3 +1,4 @@
+export * from './jobs';
 export * from './packages';
 export * from './plan';
 export * from './setup';
