+++ src/core/tasks/impl/abstract.task.ts
@@ -1,6 +1,7 @@
 import { mapIt } from '@proc7ts/a-iterable';
+import type { ZTaskExecution } from '../../jobs';
 import type { ZPackage, ZPackageSet } from '../../packages';
-import type { ZCall, ZCallDetails, ZCallPlanner, ZPrePlanner, ZTaskExecution, ZTaskParams } from '../../plan';
+import type { ZCall, ZCallDetails, ZCallPlanner, ZPrePlanner, ZTaskParams } from '../../plan';
 import type { ZTask, ZTaskQualifier } from '../task';
 import type { ZTaskBuilder$ } from '../task-builder.impl';
 import type { ZTaskSpec } from '../task-spec';
